requirements.txt
protobuf==3.20.3
google-generativeai
python-dotenv
mysql-connector-python
python-multipart
deepgram-sdk
fastapi
uvicorn
requests
plotly
Routes
kaleido